Consolidate script.
